Task 2 of seminar 1 added

+++ Java_HW/src/main/java/gd/homework/Task1.java
@@ -33,8 +33,8 @@ public abstract class Task1 {
         return result;
     }
 
-    public static int factorial() {
-        int result = 1;
+    public static float factorial() {
+        float result = 1f;
         for (int i = 1; i <= n; i++) {
             result *= i;
         }
